% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{bmcm_performance_measures_N}
\alias{bmcm_performance_measures_N}
\title{statistics for all endpoints
for full probabilistic analysis}
\usage{
bmcm_performance_measures_N(stan_out_list, par_nm, true_vals, append = TRUE)
}
\description{
statistics for all endpoints
for full probabilistic analysis
}
