% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{performance_measures_cluster}
\alias{performance_measures_cluster}
\title{Performance measures from cluster}
\usage{
performance_measures_cluster(stan_out_list, par_nm, true_vals, append = TRUE)
}
\arguments{
\item{stan_out_list}{}

\item{par_nm}{}

\item{true_vals}{}

\item{append}{when there are multiple curves near to add a number after par_nm}
}
\description{
statistics for all endpoints for full probabilistic analysis
using posterior samples obtain using cluster
}
