% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv}
\alias{rsurv}
\title{Generate time to event sample with censoring}
\usage{
rsurv(n = 100, distn = "exp", prop_cens = 0.1, params = list(rate = 1))
}
\description{
Generate time to event sample with censoring
}
