% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv_cf}
\alias{rsurv_cf}
\title{Survival mixture cure fraction model simulation
for deterministic cure fraction and full
random sampling}
\usage{
rsurv_cf(
  nsample = 20,
  n_endpoints = 2,
  t_cutpoint = 5,
  mu_cf = 0.2,
  sigma_cf,
  cf_sample_method = "random",
  cf_indiv = "fixed",
  distn = "exp",
  prop_cens = 0,
  params = list(list(rate = 1), list(rate = 1))
)
}
\description{
Survival mixture cure fraction model simulation
for deterministic cure fraction and full
random sampling
}
