% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv_mix}
\alias{rsurv_mix}
\title{Survival mixture cure fraction model simulation}
\usage{
rsurv_mix(
  nsample = 20,
  n_endpoints = 2,
  t_cutpoint,
  mu_cf = 0.2,
  sigma_cf,
  cf_sample_method = "random",
  distn = "exp",
  prop_cens = 0,
  params = list(list(rate = 1), list(rate = 1))
)
}
\description{
Survival mixture cure fraction model simulation
}
