% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run scenario}
\usage{
run_scenario(
  x,
  sim_params,
  bmcm_params,
  dir = "",
  seed = NULL,
  rstan_format = FALSE
)
}
\arguments{
\item{x}{sample number}

\item{rstan_format}{list (rstan) or flat dataframe (cmdstanr) format; default FALSE}
}
\description{
obtain posterior samples for a single simulated data set
}
