% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{bmcm_performance_measures}
\alias{bmcm_performance_measures}
\title{statistics for all endpoints}
\usage{
bmcm_performance_measures(fit, par_nm, true_vals)
}
\description{
statistics for all endpoints
}
